New GSI (ingest-id-index) for tracking ingests.
New GSI for the Dynamo S3 index that tracks ingests by collection,
experiment, and channel ids as well as resolution and job id.  The GSI
also has an int appended to it to avoid hot partitions during ingest.

The old GSI is restored but no cuboids from new ingests will be tracked
by it.  The old GSI will be used when old channels are deleted.